move MapSearch to SearchBar
