+++ frontend/src/components/cards/SearchBar.vue
